+++ src/main/kotlin/server/db/mongo/entities/CardEntity.kt
@@ -4,6 +4,6 @@ import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.mapping.Document
 
 @Document
-class CardEntity(@Id val id: Long, val name: String, val type: String, val accountId: Int): AnyEntity {
+class CardEntity(@Id val id: Long, val name: String, val type: String, val accountId: Long): AnyEntity {
     override fun toString() = "CardEntity(id=$id, name=$name, type=$type, accountId=$accountId)"
 }
